@@ -14,8 +14,11 @@
     "prepublishOnly": "npm run build"
   },
   "dependencies": {
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.11.2",
+    "@emotion/react": "^11.10.4",
+    "@emotion/styled": "^11.10.4",
+    "@mui/icons-material": "^5.10.9",
+    "@mui/material": "^5.10.10",
+    "@mui/styles": "^5.10.10",
     "react-responsive-carousel": "^3.2.23"
   },
   "devDependencies": {
@@ -24,9 +27,8 @@
     "@babel/preset-react": "^7.12.1",
     "babel-jest": "^26.6.1",
     "jest": "^26.6.1",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-scripts": "^4.0.3",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "rollup": "^2.32.1",
     "rollup-plugin-babel": "^4.4.0",
     "standard": "^15.0.0"
