@@ -1,10 +1,6 @@
 import React, { useState, useRef, useEffect } from 'react'
-import { KeyboardArrowDownOutlined } from '@material-ui/icons'
-import license from './license'
+import { KeyboardArrowDownOutlined } from '@mui/icons-material'
 import {
-  Dialog,
-  DialogTitle,
-  DialogActions,
   Typography,
   ButtonGroup,
   Button,
@@ -14,62 +10,53 @@ import {
   ClickAwayListener,
   MenuList,
   MenuItem
-} from '@material-ui/core'
-import { makeStyles } from '@material-ui/core/styles'
+} from '@mui/material'
+import makeStyles from '@mui/styles/makeStyles';
+import OpenInNew from '@mui/icons-material/OpenInNew'
 
 const useStyles = makeStyles(theme => ({
-  buttonStyle: {
-
-  },
-  pre: {
-    wordWrap: 'break-word',
-    fontFamily: 'Times New Roman',
-    padding: theme.spacing(2),
-    boxSizing: 'border-box',
-    overflowX: 'auto',
-    overflowY: 'scroll',
-    whiteSpace: 'pre-wrap',
-    maxHeight: '30em'
+  button: {
+    textTransform: 'none !important'
   }
 }), {
   name: 'BabbDownloadButton'
 })
 const options = [
   {
-    buttonText: 'Download for Mac',
+    buttonText: 'Install (macOS)',
     downloadFilename: 'Babbage Desktop.dmg',
     downloadURL: '/desktop/res/Babbage%20Desktop.dmg'
   },
   {
     buttonText:
-      'Download for Windows',
+      'Install (Windows)',
     downloadFilename: 'Babbage Desktop.exe',
     downloadURL: '/desktop/res/Babbage%20Desktop.exe'
   },
   {
     buttonText:
-      'Download for Linux (snap)',
+      'Install (Linux snap)',
     downloadFilename: 'Babbage Desktop.snap',
     downloadURL: '/desktop/res/Babbage%20Desktop.snap'
   },
   {
     buttonText:
-      'Download for Linux (AppImage)',
+      'Install (Linux AppImage)',
     downloadFilename: 'Babbage Desktop.AppImage',
     downloadURL: '/desktop/res/Babbage%20Desktop.AppImage'
   },
   {
     buttonText:
-      'Babbage Mobile (coming soon)'
+      'Stageline & Devline'
   }
 ]
 
-export default function SplitButton () {
-  const [dialogOpen, setDialogOpen] = useState(false)
-  const [open, setOpen] = useState(false)
+export default function SplitButton ({ googlePlayLink, appStoreLink }) {
+  const [menuOpen, setMenuOpen] = useState(false)
   const [downloadURL, setDownloadURL] = useState('/desktop/res/Babbage%20Desktop.exe')
   const anchorRef = useRef(null)
   const [selectedOS, setSelectedOS] = useState(1)
+  const [downloading, setDownloading] = useState(false)
   const classes = useStyles()
   const [downloadFilename, setDownloadFilename] = useState(
     'Babbage Desktop.exe'
@@ -77,18 +64,17 @@ export default function SplitButton () {
   const [disabled, setDisabled] = useState(false)
   const handleDownload = () => {
     const a = document.createElement('a')
-    a.href = `${downloadURL}`
+    a.href = `https://projectbabbage.com${downloadURL}`
     a.download = downloadFilename
     a.click()
-    setDialogOpen(false)
-  }
-  const handleClick = () => {
-    console.info(`You clicked ${options[selectedOS]}`)
-    setDialogOpen(true)
+    setDownloading(true)
+    setTimeout(() => {
+      setDownloading(false)
+    }, 3500)
   }
 
   const handleToggle = () => {
-    setOpen((prevOpen) => !prevOpen)
+    setMenuOpen((prevOpen) => !prevOpen)
   }
   useEffect(() => {
     // Thanks to https://stackoverflow.com/a/38241481 for this
@@ -114,8 +100,7 @@ export default function SplitButton () {
     if (anchorRef.current && anchorRef.current.contains(event.target)) {
       return
     }
-
-    setOpen(false)
+    setMenuOpen(false)
   }
   if (disabled) {
     return (
@@ -126,37 +111,38 @@ export default function SplitButton () {
   }
   return (
     <>
-      <ButtonGroup className={classes.buttonStyle} size='small' color='secondary' variant='contained' ref={anchorRef} aria-label='split button'>
-        <Button onClick={handleClick}>{options[selectedOS].buttonText}</Button>
+      <ButtonGroup
+        className={classes.buttonStyle}
+        size='small'
+        color='primary'
+        variant='contained'
+        ref={anchorRef}
+        aria-label='split button'
+        disabled={downloading}
+      >
+        <Button
+          className={classes.button}
+          onClick={handleDownload}
+        >
+          {downloading ? 'Downloading...' : options[selectedOS].buttonText}
+        </Button>
         <Button
           size='small'
-          aria-controls={open ? 'split-button-menu' : undefined}
-          aria-expanded={open ? 'true' : undefined}
-          aria-label='select merge strategy'
+          aria-controls={menuOpen ? 'split-button-menu' : undefined}
+          aria-expanded={menuOpen ? 'true' : undefined}
+          aria-label='select desktop build to install'
           aria-haspopup='menu'
+          color='secondary'
           onClick={handleToggle}
         >
           <KeyboardArrowDownOutlined />
         </Button>
-        <Dialog
-          open={dialogOpen}
-          onClose={() => setDialogOpen(false)}
-        >
-          <DialogTitle>Babbage Desktop License Agreement</DialogTitle>
-          <pre className={classes.pre}>
-            {license}
-          </pre>
-          <DialogActions>
-            <Button onClick={() => setDialogOpen(false)}>I Disagree</Button>
-            <Button onClick={handleDownload}>I Agree</Button>
-          </DialogActions>
-        </Dialog>
       </ButtonGroup>
       <Popper
         sx={{
           zIndex: 1
         }}
-        open={open}
+        open={menuOpen}
         anchorEl={anchorRef.current}
         role={undefined}
         transition
@@ -176,16 +162,22 @@ export default function SplitButton () {
                   {options.map((option, index) => (
                     <MenuItem
                       key={index}
-                      disabled={index === 4}
                       selected={index === selectedOS}
                       onClick={(event) => {
-                        setDownloadURL(option.downloadURL)
-                        setDownloadFilename(option.downloadFilename)
-                        setSelectedOS(index)
-                        setOpen(false)
+                        if (index === 4) {
+                          window.open(
+                            'https://projectbabbage.com/docs/dev-downloads',
+                            '_blank'
+                          )
+                        } else {
+                          setDownloadURL(option.downloadURL)
+                          setDownloadFilename(option.downloadFilename)
+                          setSelectedOS(index)
+                        }
+                        setMenuOpen(false)
                       }}
                     >
-                      {option.buttonText}
+                      {option.buttonText} {index === 4 && <OpenInNew />}
                     </MenuItem>
                   ))}
                 </MenuList>
